Vaguely Working
Vaguely Working - Switch back to byte array everywhere for speed

+++ DCS-SR-Client/Audio/ClientAudio.cs
@@ -5,7 +5,7 @@ namespace Ciribob.DCS.SimpleRadio.Standalone.Client
 {
     public class ClientAudio
     {
-        public byte[] PcmAudioFloat { get; set; }
+        public float[] PcmAudioFloat { get; set; }
         public string ClientGuid { get; set; }
         public long ReceiveTime { get; set; }
         public int ReceivedRadio { get; set; }
